refactor(solver): split evaluator support methods
